Separate splash level
